LPS-159400 Move class SalesforceConfiguration to api module
