+++ modules/apps/object/object-storage-salesforce-api/bnd.bnd
@@ -1,0 +1,4 @@
+Bundle-Name: Liferay Object Storage Salesforce API
+Bundle-SymbolicName: com.liferay.object.storage.salesforce.api
+Bundle-Version: 1.0.0
+Export-Package: com.liferay.object.storage.salesforce.configuration
