+++ modules/apps/object/object-storage-salesforce-api/build.gradle
@@ -1,0 +1,4 @@
+dependencies {
+	compileOnly group: "com.liferay", name: "biz.aQute.bnd.annotation", version: "4.2.0.LIFERAY-PATCHED-2"
+	compileOnly project(":apps:static:portal-configuration:portal-configuration-metatype-api")
+}
