+++ modules/apps/object/object-storage-salesforce-api/src/main/resources/com/liferay/object/storage/salesforce/configuration/packageinfo
@@ -1,0 +1,1 @@
+version 1.0.0
